refactor: modularize products translations
