--- packages/plugin-commerce-dashboard/src/locales/en/products.json
@@ -1,117 +1,0 @@
-{
-  "products": {
-    "pageTitle": "Products",
-    "pageDescription": "Manage your catalogue, pricing and availability.",
-    "actions": {
-      "refresh": "Refresh",
-      "create": "New product"
-    },
-    "buttons": {
-      "add": "Add product",
-      "copy": "Copy",
-      "download": "Download",
-      "edit": "Edit",
-      "delete": "Delete"
-    },
-    "search": {
-      "placeholder": "Search products...",
-      "minLength": "Type at least 3 characters to search."
-    },
-    "filters": {
-      "active": "{{count}} active filters",
-      "activeView": "View: {{name}}",
-      "clear": "Clear filters",
-      "fields": {
-        "status": "Status",
-        "tags": "Tags",
-        "collection": "Collection ID",
-        "publishAt": "Publish at",
-        "expireAt": "Unpublish at",
-        "createdAt": "Created at"
-      },
-      "views": {
-        "draft": {
-          "description": "Show products that are still in draft"
-        },
-        "active": {
-          "description": "Show products that are live"
-        },
-        "archived": {
-          "description": "Show products that are archived"
-        }
-      },
-      "toasts": {
-        "viewSaved": {
-          "title": "View saved",
-          "description": "Saved \"{{name}}\" for quick reuse."
-        },
-        "viewDeleted": {
-          "title": "View deleted",
-          "description": "Removed \"{{name}}\" from your saved filters."
-        },
-        "viewSaveError": "Unable to save view. Please try again.",
-        "viewDeleteError": "Unable to delete view. Please try again."
-      }
-    },
-    "emptyState": {
-      "title": "No products yet",
-      "description": "Start building your catalogue by creating your first product."
-    },
-    "table": {
-      "error": "Unable to load products. Please retry.",
-      "columns": {
-        "name": "Name",
-        "status": "Status",
-        "currency": "Currency",
-        "updatedAt": "Last updated"
-      },
-      "untitled": "Untitled product",
-      "empty": "No products to display"
-    },
-    "fields": {
-      "title": "Title",
-      "status": "Status",
-      "languages": "Languages",
-      "tags": "Tags",
-      "currency": "Currency",
-      "publishAt": "Publish at",
-      "updatedAt": "Updated at",
-      "actions": "Actions"
-    },
-    "status": {
-      "draft": "Draft",
-      "active": "Active",
-      "archived": "Archived"
-    },
-    "details": {
-      "title": "Product details",
-      "description": "Review the product configuration and localisation.",
-      "breadcrumb": "Product details",
-      "generalInformation": "General information",
-      "status": "Status",
-      "currency": "Default currency",
-      "publishAt": "Scheduled publish",
-      "expireAt": "Scheduled expiration",
-      "contentSection": "Content",
-      "noSummary": "No summary available for this language.",
-      "slug": "Slug",
-      "additionalInformation": "Additional information",
-      "tags": "Tags",
-      "media": "Media",
-      "galleryCount": "{{count}} assets in gallery",
-      "updatedAt": "Last updated"
-    },
-    "delete": {
-      "title": "Delete product",
-      "description": "Are you sure you want to delete this product? This action cannot be undone.",
-      "cancel": "Cancel",
-      "confirm": "Delete"
-    },
-    "create": {
-      "title": "Create product",
-      "description": "The visual editor for products will be available soon.",
-      "helper": "In the meantime you can create products via the API or integrate your own editor interface.",
-      "breadcrumb": "Create product"
-    }
-  }
-}
