+++ packages/plugin-commerce-dashboard/src/locales/en/products/actions.json
@@ -1,0 +1,4 @@
+{
+  "refresh": "Refresh",
+  "create": "New product"
+}
