+++ packages/plugin-commerce-dashboard/src/locales/en/products/buttons.json
@@ -1,0 +1,7 @@
+{
+  "add": "Add product",
+  "copy": "Copy",
+  "download": "Download",
+  "edit": "Edit",
+  "delete": "Delete"
+}
