+++ packages/plugin-commerce-dashboard/src/locales/en/products/common.json
@@ -1,0 +1,4 @@
+{
+  "pageTitle": "Products",
+  "pageDescription": "Manage your catalogue, pricing and availability."
+}
