+++ packages/plugin-commerce-dashboard/src/locales/en/products/create.json
@@ -1,0 +1,6 @@
+{
+  "title": "Create product",
+  "description": "The visual editor for products will be available soon.",
+  "helper": "In the meantime you can create products via the API or integrate your own editor interface.",
+  "breadcrumb": "Create product"
+}
