+++ packages/plugin-commerce-dashboard/src/locales/en/products/empty-state.json
@@ -1,0 +1,4 @@
+{
+  "title": "No products yet",
+  "description": "Start building your catalogue by creating your first product."
+}
