+++ packages/plugin-commerce-dashboard/src/locales/en/products/search.json
@@ -1,0 +1,4 @@
+{
+  "placeholder": "Search products...",
+  "minLength": "Type at least 3 characters to search."
+}
