+++ packages/plugin-commerce-dashboard/src/locales/en/products/status.json
@@ -1,0 +1,5 @@
+{
+  "draft": "Draft",
+  "active": "Active",
+  "archived": "Archived"
+}
