+++ packages/plugin-commerce-dashboard/src/locales/en/products/table.json
@@ -1,0 +1,11 @@
+{
+  "error": "Unable to load products. Please retry.",
+  "columns": {
+    "name": "Name",
+    "status": "Status",
+    "currency": "Currency",
+    "updatedAt": "Last updated"
+  },
+  "untitled": "Untitled product",
+  "empty": "No products to display"
+}
