+++ packages/plugin-commerce-dashboard/src/locales/it/index.ts
@@ -1,6 +1,6 @@
 import shared from "./shared.json";
 import settings from "./settings.json";
-import products from "./products.json";
+import products from "./products";
 import orders from "./orders.json";
 import collections from "./collections";
