--- packages/plugin-commerce-dashboard/src/locales/it/products.json
@@ -1,117 +1,0 @@
-{
-  "products": {
-    "pageTitle": "Prodotti",
-    "pageDescription": "Gestisci catalogo, prezzi e disponibilità.",
-    "actions": {
-      "refresh": "Aggiorna",
-      "create": "Nuovo prodotto"
-    },
-    "buttons": {
-      "add": "Aggiungi prodotto",
-      "copy": "Copia",
-      "download": "Scarica",
-      "edit": "Modifica",
-      "delete": "Elimina"
-    },
-    "search": {
-      "placeholder": "Cerca prodotti...",
-      "minLength": "Inserisci almeno 3 caratteri per cercare."
-    },
-    "filters": {
-      "active": "{{count}} filtri attivi",
-      "activeView": "Vista: {{name}}",
-      "clear": "Pulisci filtri",
-      "fields": {
-        "status": "Stato",
-        "tags": "Tag",
-        "collection": "ID collezione",
-        "publishAt": "Pubblica il",
-        "expireAt": "Depubblica il",
-        "createdAt": "Creato il"
-      },
-      "views": {
-        "draft": {
-          "description": "Mostra i prodotti ancora in bozza"
-        },
-        "active": {
-          "description": "Mostra i prodotti già online"
-        },
-        "archived": {
-          "description": "Mostra i prodotti archiviati"
-        }
-      },
-      "toasts": {
-        "viewSaved": {
-          "title": "Vista salvata",
-          "description": "\"{{name}}\" è stata salvata per un riutilizzo rapido."
-        },
-        "viewDeleted": {
-          "title": "Vista eliminata",
-          "description": "\"{{name}}\" è stata rimossa dalle viste salvate."
-        },
-        "viewSaveError": "Impossibile salvare la vista. Riprova.",
-        "viewDeleteError": "Impossibile eliminare la vista. Riprova."
-      }
-    },
-    "emptyState": {
-      "title": "Nessun prodotto",
-      "description": "Inizia a costruire il catalogo creando il primo prodotto."
-    },
-    "table": {
-      "error": "Impossibile caricare i prodotti. Riprova.",
-      "columns": {
-        "name": "Nome",
-        "status": "Stato",
-        "currency": "Valuta",
-        "updatedAt": "Ultimo aggiornamento"
-      },
-      "untitled": "Prodotto senza titolo",
-      "empty": "Nessun prodotto da mostrare"
-    },
-    "fields": {
-      "title": "Titolo",
-      "status": "Stato",
-      "languages": "Lingue",
-      "tags": "Tag",
-      "currency": "Valuta",
-      "publishAt": "Pubblica il",
-      "updatedAt": "Aggiornato il",
-      "actions": "Azioni"
-    },
-    "status": {
-      "draft": "Bozza",
-      "active": "Attivo",
-      "archived": "Archiviato"
-    },
-    "details": {
-      "title": "Dettagli prodotto",
-      "description": "Consulta configurazione e localizzazioni del prodotto.",
-      "breadcrumb": "Dettagli prodotto",
-      "generalInformation": "Informazioni generali",
-      "status": "Stato",
-      "currency": "Valuta predefinita",
-      "publishAt": "Pubblicazione programmata",
-      "expireAt": "Scadenza programmata",
-      "contentSection": "Contenuti",
-      "noSummary": "Nessun riassunto disponibile per questa lingua.",
-      "slug": "Slug",
-      "additionalInformation": "Informazioni aggiuntive",
-      "tags": "Tag",
-      "media": "Media",
-      "galleryCount": "{{count}} elementi in galleria",
-      "updatedAt": "Ultimo aggiornamento"
-    },
-    "delete": {
-      "title": "Elimina prodotto",
-      "description": "Sei sicuro di voler eliminare questo prodotto? L'azione è irreversibile.",
-      "cancel": "Annulla",
-      "confirm": "Elimina"
-    },
-    "create": {
-      "title": "Crea prodotto",
-      "description": "L'editor visuale per i prodotti sarà presto disponibile.",
-      "helper": "Nel frattempo puoi creare prodotti via API o integrare una tua interfaccia.",
-      "breadcrumb": "Crea prodotto"
-    }
-  }
-}
