+++ packages/plugin-commerce-dashboard/src/locales/it/products/actions.json
@@ -1,0 +1,4 @@
+{
+  "refresh": "Aggiorna",
+  "create": "Nuovo prodotto"
+}
