+++ packages/plugin-commerce-dashboard/src/locales/it/products/buttons.json
@@ -1,0 +1,7 @@
+{
+  "add": "Aggiungi prodotto",
+  "copy": "Copia",
+  "download": "Scarica",
+  "edit": "Modifica",
+  "delete": "Elimina"
+}
