+++ packages/plugin-commerce-dashboard/src/locales/it/products/common.json
@@ -1,0 +1,4 @@
+{
+  "pageTitle": "Prodotti",
+  "pageDescription": "Gestisci catalogo, prezzi e disponibilità."
+}
