+++ packages/plugin-commerce-dashboard/src/locales/it/products/create.json
@@ -1,0 +1,6 @@
+{
+  "title": "Crea prodotto",
+  "description": "L'editor visuale per i prodotti sarà presto disponibile.",
+  "helper": "Nel frattempo puoi creare prodotti via API o integrare una tua interfaccia.",
+  "breadcrumb": "Crea prodotto"
+}
