+++ packages/plugin-commerce-dashboard/src/locales/it/products/delete.json
@@ -1,0 +1,6 @@
+{
+  "title": "Elimina prodotto",
+  "description": "Sei sicuro di voler eliminare questo prodotto? L'azione è irreversibile.",
+  "cancel": "Annulla",
+  "confirm": "Elimina"
+}
