+++ packages/plugin-commerce-dashboard/src/locales/it/products/empty-state.json
@@ -1,0 +1,4 @@
+{
+  "title": "Nessun prodotto",
+  "description": "Inizia a costruire il catalogo creando il primo prodotto."
+}
