+++ packages/plugin-commerce-dashboard/src/locales/it/products/search.json
@@ -1,0 +1,4 @@
+{
+  "placeholder": "Cerca prodotti...",
+  "minLength": "Inserisci almeno 3 caratteri per cercare."
+}
