+++ packages/plugin-commerce-dashboard/src/locales/it/products/status.json
@@ -1,0 +1,5 @@
+{
+  "draft": "Bozza",
+  "active": "Attivo",
+  "archived": "Archiviato"
+}
